# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SourceAdminOut": {
        "properties": {
          "channel_key": {
            "title": "Channel Key",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "last_listener_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Listener Error"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "spotter",
              "alert"
            ],
            "title": "Role",
            "type": "string"
          },
          "stats": {
            "$ref": "#/components/schemas/SourceStatsOut"
          },
          "subscribe_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscribe Ref"
          },
          "trust_weight": {
            "title": "Trust Weight",
            "type": "number"
          }
        },
        "required": [
          "id",
          "channel_key",
          "name",
          "subscribe_ref",
          "role",
          "is_active",
          "trust_weight",
          "last_listener_error",
          "created_at",
          "stats"
        ],
        "title": "SourceAdminOut",
        "type": "object"
      },
      "SourceDeleteOut": {
        "description": "What a hard-delete removed — surfaced so the admin sees the blast radius.",
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "incidents_deleted": {
            "title": "Incidents Deleted",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notices": {
            "title": "Notices",
            "type": "integer"
          },
          "raw_messages": {
            "title": "Raw Messages",
            "type": "integer"
          },
          "threats_deleted": {
            "title": "Threats Deleted",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "raw_messages",
          "events",
          "notices",
          "threats_deleted",
          "incidents_deleted"
        ],
        "title": "SourceDeleteOut",
        "type": "object"
      },
      "SourceIn": {
        "description": "Add (or reactivate) a channel from the admin console.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "role": {
            "default": "spotter",
            "enum": [
              "spotter",
              "alert"
            ],
            "title": "Role",
            "type": "string"
          },
          "subscribe_ref": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Subscribe Ref",
            "type": "string"
          },
          "trust_weight": {
            "default": 1.0,
            "title": "Trust Weight",
            "type": "number"
          }
        },
        "required": [
          "subscribe_ref"
        ],
        "title": "SourceIn",
        "type": "object"
      },
      "SourceStatsOut": {
        "description": "Per-source quality signals, computed on-demand (JOIN by source_id) over a\nrecent window — see app/api/source_stats.py. Rates are None when there's no\ndenominator (e.g. llm_fallback_rate before any row had llm_attempted set).",
        "properties": {
          "conflict_share": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conflict Share"
          },
          "correction_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction Rate"
          },
          "coverage_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Rate"
          },
          "events_produced": {
            "title": "Events Produced",
            "type": "integer"
          },
          "last_message_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Message At"
          },
          "llm_fallback_rate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Fallback Rate"
          },
          "messages_processed": {
            "title": "Messages Processed",
            "type": "integer"
          },
          "messages_total": {
            "title": "Messages Total",
            "type": "integer"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          }
        },
        "required": [
          "messages_total",
          "messages_processed",
          "events_produced",
          "llm_fallback_rate",
          "coverage_rate",
          "correction_rate",
          "conflict_share",
          "quality_score",
          "last_message_at"
        ],
        "title": "SourceStatsOut",
        "type": "object"
      },
      "SourceUpdateIn": {
        "properties": {
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "role": {
            "anyOf": [
              {
                "enum": [
                  "spotter",
                  "alert"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "trust_weight": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trust Weight"
          }
        },
        "title": "SourceUpdateIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Kyiv Aerial Threat Tracker (MVP)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/sources": {
      "get": {
        "description": "Every source with its on-demand quality stats — active channels first,\nthen by name. Inactive rows are kept (removal is a soft deactivate) so the\noperator can re-enable or inspect a channel's history.",
        "operationId": "admin_list_sources_admin_sources_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SourceAdminOut"
                  },
                  "title": "Response Admin List Sources Admin Sources Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin List Sources"
      },
      "post": {
        "description": "Add (or reactivate) a channel and tell the listener to re-subscribe. Whether\nit actually starts delivering shows up as message counts / last_listener_error\non the next connect.",
        "operationId": "admin_add_source_admin_sources_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Add Source"
      }
    },
    "/admin/sources/{source_id}": {
      "delete": {
        "description": "HARD-delete a channel AND all of its stored data (raw_messages, notices,\nthreat_events), repairing the tracks/incidents it touched — irreversible,\nunlike deactivate. Refused while an attack is active (the graph surgery would\ndisrupt live tracking) unless `force`.",
        "operationId": "admin_delete_source_admin_sources__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "description": "Delete even while an attack is active",
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "description": "Delete even while an attack is active",
              "title": "Force",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceDeleteOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Delete Source"
      },
      "patch": {
        "description": "Edit a source's name / role / trust_weight / active flag. Only a change that\naffects what's watched (role or is_active) triggers a listener reload.",
        "operationId": "admin_update_source_admin_sources__source_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SourceUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Update Source"
      }
    },
    "/admin/sources/{source_id}/activate": {
      "post": {
        "operationId": "admin_activate_source_admin_sources__source_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Activate Source"
      }
    },
    "/admin/sources/{source_id}/deactivate": {
      "post": {
        "description": "Remove a channel from the live feed (soft — the row stays, since\nraw_messages/threat_events reference it and its history stays queryable).",
        "operationId": "admin_deactivate_source_admin_sources__source_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SourceAdminOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Deactivate Source"
      }
    },
...
  }
}